Fixed SearchDetail, AppDelegate, MapVC, and SearchVc. Deleted SettingsVC and SettingsView
