+++ Venues_App/Controller/MapViewController.swift
@@ -164,25 +164,5 @@ extension MapViewController: MKMapViewDelegate {
         mapView.mapView.removeOverlays(mapView.mapView.overlays)
         directions.append(newDirections)
         let _ = directions.map { $0.cancel() }
-
-        // TODO: need to safe unwrap coordinate
-        getDirections(venueCoordinate: view.annotation!.coordinate)
-    }
-    
-    private func getDirections(venueCoordinate: CLLocationCoordinate2D) {
-        guard let userLocation = locationManager.location?.coordinate else {
-            showAlert(title: "Error Message", message: "Warning we can't seem to find your location!")
-            return
-        }
-//        let request = createDirectionsRequest(destinationCoordinate: venueCoordinate, from: userLocation)
-    }
-    
-//    private func createDirectionsRequest(destinationCoordinate: CLLocationCoordinate2D, from coordinate: CLLocationCoordinate2D) -> MKDirections.Request {
-//
-//    }
-    
-    private func resetMapView(withNew directions: MKDirections) {
-        
-
     }
 }
