+++ Venues_App/View/SearchView.swift
@@ -24,7 +24,17 @@ class SearchView: UIView {
         searchBar.delegate = self
         searchBar.showsCancelButton = true
         searchBar.searchBarStyle = UISearchBar.Style.default
-        searchBar.placeholder = " Search Here....."
+        searchBar.placeholder = " Search Location or Place i.e. Miami or Queens"
+        searchBar.sizeToFit()
+        return searchBar
+    }()
+    
+    lazy var defaultLocationSearchBar: UISearchBar = {
+        let searchBar = UISearchBar()
+        searchBar.frame = CGRect(x: 0, y: 0, width: 200, height: 70)
+        searchBar.delegate = self
+        searchBar.showsCancelButton = true
+        searchBar.placeholder = "  Type in Venue i.e. Coffee, Bakery, Food"
         searchBar.sizeToFit()
         return searchBar
     }()
@@ -49,6 +59,7 @@ class SearchView: UIView {
     private func commonInit() {
         setupSearchBar()
         setupTableView()
+        setupDefaultLocationSearchBar()
         
     }
     
@@ -58,6 +69,15 @@ class SearchView: UIView {
         searchBar.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: 0).isActive = true
         searchBar.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor, constant: 0).isActive = true
         searchBar.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: 0).isActive = true
+    }
+    
+    private func setupDefaultLocationSearchBar() {
+        addSubview(defaultLocationSearchBar)
+        defaultLocationSearchBar.translatesAutoresizingMaskIntoConstraints = false
+        defaultLocationSearchBar.topAnchor.constraint(equalTo: searchBar.bottomAnchor, constant: 0).isActive = true
+        defaultLocationSearchBar.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 0).isActive = true
+        defaultLocationSearchBar.trailingAnchor.constraint(equalTo: trailingAnchor, constant: 0).isActive = true
+        defaultLocationSearchBar.heightAnchor.constraint(equalToConstant: 60).isActive = true
     }
     
     private func setupTableView() {
